+++ app/src/main/java/woowacourse/shopping/ui/productdetail/ProductDetailViewModel.kt
@@ -8,11 +8,11 @@ import kotlinx.coroutines.flow.asStateFlow
 import kotlinx.coroutines.flow.update
 import kotlinx.coroutines.launch
 import woowacourse.shopping.model.ProductId
+import woowacourse.shopping.network.NetworkMonitor
 import woowacourse.shopping.repository.CartRepository
 import woowacourse.shopping.repository.ProductRepository
 import woowacourse.shopping.repository.RecentProductRepository
 import woowacourse.shopping.repository.ShoppingRepositoryProvider
-import woowacourse.shopping.repository.network.NetworkMonitor
 import woowacourse.shopping.ui.common.recentlyviewed.RecentViewedProductsMapper
 
 class ProductDetailViewModel(
